Added Accordion handling to the menu.  Also added a section for the sayings that KIPP has.  All the links will have to be setup however.
At the bequest of Karthik, I've added a message section under the menu as well.  This area will be utilized later for timed messages like: Informing the Administrator that the Technical Admin must reset the system for the next school year.  It will of course have to be programmed, provided we have the time to implement.